+++ drizzle-config.shared.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'drizzle-kit';
+
+export function createDrizzleConfig(url: string) {
+  return defineConfig({
+    out: './drizzle',
+    schema: './src/db/schema.ts',
+    dialect: 'postgresql',
+    dbCredentials: {
+      url,
+    },
+  });
+}
