+++ drizzle-prod.config.ts
@@ -1,4 +1,4 @@
-import { createDrizzleConfig } from './drizzle.config';
+import { createDrizzleConfig } from './drizzle-config.shared';
 
 if (!process.env.PROD_DATABASE_URL) {
   throw new Error('PROD_DATABASE_URL is required for prod Drizzle operations');
