+++ drizzle.config.ts
@@ -1,20 +1,9 @@
 /// <reference types="node" />
 
-import { defineConfig } from 'drizzle-kit';
+import { createDrizzleConfig } from './drizzle-config.shared';
 
 if (!process.env.DATABASE_URL) {
   throw new Error('DATABASE_URL is required for Drizzle operations');
-}
-
-export function createDrizzleConfig(url: string) {
-  return defineConfig({
-    out: './drizzle',
-    schema: './src/db/schema.ts',
-    dialect: 'postgresql',
-    dbCredentials: {
-      url,
-    },
-  });
 }
 
 export default createDrizzleConfig(process.env.DATABASE_URL);
